requirements.txt
aiohttp==3.10.5
discord.py==2.4.0
python-dotenv==1.0.1
Flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
